requirements.txt
# COMP90024 Project - Team 34
# Lokesh Sai Sri Harsha Sankarasetty, Melbourne, [1130612]
# Kanch Vatcharotayan, Melbourne, [1132855]
# Sai Deepthi Amancha, Melbourne, [1051388]
# Josin Saji Abraham, Melbourne, [1129428]
# Kush Garg, Melbourne, [1146696]

starlette
pydantic==1.5.1
nltk==3.5
fastapi==0.54.1
uvicorn==0.11.5
numpy==1.18.4
tweepy==3.8.0
cloudant==2.13.0
shapely==1.7.0
textblob==0.15.3
python-dotenv==0.13.0
google-cloud-translate==2.0.1